converted KeySearch into two smaller components
